Never emit the CRC when armoring
This might be too abrupt of a change.  The crypto-refresh says to
*never* emit the CRC if there's a v5 secret key or a v2 SEIPD.  This
complies with that, but might produce v4 data that legacy
implementations can't handle.

Probably better would be to make an Armorable class be able to report
whether to include the CRC or not, and emit it as needed.

It will also break any test that relies on deterministic armored
output.