S2KSpecifier: speed up derive_key
Rather than invoking the cryptography module's hash update function a
bunch of times, we'll chew up up to 64MiB RAM at once (this is the max
iteration octets) and invoke the hash update function only once.

This is the same insight that Michael Greene had in
280d82c74a7170113803b75648c5885ede00b07a, but i'd missed during the
S2KSpecifier breakout.

On my hardware, this makes "sopgpy encrypt --with-password" over a
small cleartextgo from ~16.5s of CPU time to ~2.5s of CPU time.