OPSv3: get the "nested" flag right
"nested" semantically probably is meant to mean "another OPS packet follows".
But the byte on the wire is defined as 0 means another OPS packet follows.

Furthermore, the flags are set in the wrong way: before this commit, the
code produced a series of OPS packets where the first packet had a different
value for the flag than all subsequent ones.  What we want is where all the
flags *except* the last OPS packet (corresponding to the *first* Sig packet)
are 0.

See https://gitlab.com/sequoia-pgp/openpgp-interoperability-test-suite/-/issues/84